Promote ov002 daObjLava_c, daObjFire_c and daSCoin_c to intact-object TUs
Three ov002 classes move from one-function shards to one reconstructed
translation unit each, owning both the .text run and the class's .data run
(_ZTI, _ZTS, spawn-info record(s), _ZTV):

  daObjLava_c  0x020b6d28..0x020b6e08  .data 0x021093a0..0x0210945c  5 fns
  daObjFire_c  0x020b5734..0x020b5a18  .data 0x02108edc..0x02108fb4  8 fns
  daSCoin_c    0x020f03c4..0x020f0894  .data 0x0210aff4..0x0210b0ac  9 fns

Two of them take their ROM RTTI names (the _ZTS string the ROM itself
carries is the class name): PoppingLavaBubbles -> daObjLava_c
(_ZTS at ov002:0x021093ac) and BlueFlame -> daObjFire_c (_ZTS at
ov002:0x02108ee8). The coined names' vtable rows and decl_common.h
declarations are deleted, not aliased; no row is added to any symbols.txt
beyond the rename itself. symbols/actor_renames.tsv gets one appended row per
renamed symbol. func_ov002_020b6d84 is daObjLava_c::Behavior (vtable slot 16)
and is renamed accordingly.

Each destructor is defined inline in the class body, which is what makes
mwccarm emit the cartridge's D1/D0 pair in the cartridge's order with no D2.
The factories store &_ZTV<C>[2] because each TU now emits its own vtable.

Proof, run here on this tree before promotion (see the handoff note):
  tubuild.py verify      5/5, 8/8, 9/9 MATCH, objisolate clean
  tubuild.py linkcheck   SCRATCH-DATA-VERIFIED for all three; text and data
                         ranges IDENTICAL, 106/106 modules, ROM sha equal to
                         the stock control
  rombuild.py --no-rom   106/106 exact, 0 mismatching, 23/23 data claims

This is batch 1 of the rescue of PR #2231 (cpp/promote-da-obj-lava @
9622b484a), re-derived on today's main rather than merged.
