+++ attribution.json
@@ -2961,6 +2961,28 @@
     "src/actors/dScMgLuigi_c.cpp#func_ov006_020f319c": "tangosdev",
     "src/actors/dScMgLuigi_c.cpp#func_ov006_020f31dc": "tangosdev",
     "src/actors/dScMgLuigi_c.cpp#func_ov006_020f3260": "tangosdev",
-    "src/actors/dScMgLuigi_c.cpp#d_s_mg_luigi": "tangosdev"
+    "src/actors/dScMgLuigi_c.cpp#d_s_mg_luigi": "tangosdev",
+    "src/actors/daObjLava_c.cpp#_ZN11daObjLava_cD1Ev": "andrewboudreau",
+    "src/actors/daObjLava_c.cpp#_ZN11daObjLava_cD0Ev": "andrewboudreau",
+    "src/actors/daObjLava_c.cpp#_ZN11daObjLava_c8BehaviorEv": "andrewboudreau",
+    "src/actors/daObjLava_c.cpp#_ZN11daObjLava_c13InitResourcesEv": "tangosdev",
+    "src/actors/daObjLava_c.cpp#daObjLava_c_classInit": "andrewboudreau",
+    "src/actors/daObjFire_c.cpp#_ZN11daObjFire_cD1Ev": "tangosdev",
+    "src/actors/daObjFire_c.cpp#_ZN11daObjFire_cD0Ev": "tangosdev",
+    "src/actors/daObjFire_c.cpp#_ZN11daObjFire_c8BehaviorEv": "ruspecial",
+    "src/actors/daObjFire_c.cpp#_ZN11daObjFire_c13InitResourcesEv": "tangosdev",
+    "src/actors/daObjFire_c.cpp#_ZN11daObjFire_c13OnTurnIntoEggER6Player": "andrewboudreau",
+    "src/actors/daObjFire_c.cpp#_ZN11daObjFire_c13OnYoshiTryEatEv": "andrewboudreau",
+    "src/actors/daObjFire_c.cpp#daObjFire_c_classInit_OBJ_BLUE_FIRE": "tangosdev",
+    "src/actors/daObjFire_c.cpp#daObjFire_c_classInit_OBJ_RED_FIRE": "tangosdev",
+    "src/actors/daSCoin_c.cpp#_ZN9daSCoin_cD1Ev": "tangosdev",
+    "src/actors/daSCoin_c.cpp#_ZN9daSCoin_cD0Ev": "tangosdev",
+    "src/actors/daSCoin_c.cpp#func_ov002_020f0438": "tangosdev",
+    "src/actors/daSCoin_c.cpp#func_ov002_020f051c": "andrewboudreau",
+    "src/actors/daSCoin_c.cpp#func_ov002_020f05f4": "ruspecial",
+    "src/actors/daSCoin_c.cpp#_ZN9daSCoin_c16CleanupResourcesEv": "tangosdev",
+    "src/actors/daSCoin_c.cpp#_ZN9daSCoin_c8BehaviorEv": "tangosdev",
+    "src/actors/daSCoin_c.cpp#_ZN9daSCoin_c13InitResourcesEv": "tangosdev",
+    "src/actors/daSCoin_c.cpp#daSCoin_c_classInit": "tangosdev"
   }
 }
