+++ config/arm9/overlays/ov002/symbols.txt
@@ -210,12 +210,12 @@ _ZN19daObjHatenaSwitch_c13InitResourcesEv kind:function(arm,size=0x17c) addr:0x0
 _ZN19daObjHatenaSwitch_c10HandleClsnER8dActor_c kind:function(arm,size=0x48) addr:0x020b567c
 _ZN19daObjHatenaSwitch_c17AfterClsnCallbackEP4dBgWP8dActor_cS3_ kind:function(arm,size=0x14) addr:0x020b56c4
 daObjHatenaSwitch_c_classInit kind:function(arm,size=0x5c) addr:0x020b56d8
-_ZN9BlueFlameD1Ev kind:function(arm,size=0x30) addr:0x020b5734
-_ZN9BlueFlameD0Ev kind:function(arm,size=0x44) addr:0x020b5764
-_ZN9BlueFlame8BehaviorEv kind:function(arm,size=0x198) addr:0x020b57a8
-_ZN9BlueFlame13InitResourcesEv kind:function(arm,size=0x5c) addr:0x020b5940
-_ZN9BlueFlame13OnTurnIntoEggER6Player kind:function(arm,size=0x4) addr:0x020b599c
-_ZN9BlueFlame13OnYoshiTryEatEv kind:function(arm,size=0x8) addr:0x020b59a0
+_ZN11daObjFire_cD1Ev kind:function(arm,size=0x30) addr:0x020b5734
+_ZN11daObjFire_cD0Ev kind:function(arm,size=0x44) addr:0x020b5764
+_ZN11daObjFire_c8BehaviorEv kind:function(arm,size=0x198) addr:0x020b57a8
+_ZN11daObjFire_c13InitResourcesEv kind:function(arm,size=0x5c) addr:0x020b5940
+_ZN11daObjFire_c13OnTurnIntoEggER6Player kind:function(arm,size=0x4) addr:0x020b599c
+_ZN11daObjFire_c13OnYoshiTryEatEv kind:function(arm,size=0x8) addr:0x020b59a0
 daObjFire_c_classInit_OBJ_BLUE_FIRE kind:function(arm,size=0x38) addr:0x020b59a8
 daObjFire_c_classInit_OBJ_RED_FIRE kind:function(arm,size=0x38) addr:0x020b59e0
 _ZN17daObjFloatBoard_cD0Ev kind:function(arm,size=0x58) addr:0x020b5a18
@@ -262,10 +262,10 @@ func_ov002_020b6ac8 kind:function(arm,size=0x48) addr:0x020b6ac8
 _ZN17daObjKurumajiku_c6RenderEv kind:function(arm,size=0x28) addr:0x020b6b10
 _ZN17daObjKurumajiku_c8BehaviorEv kind:function(arm,size=0x11c) addr:0x020b6b38
 func_ov002_020b6c54 kind:function(arm,size=0xd4) addr:0x020b6c54
-_ZN18PoppingLavaBubblesD1Ev kind:function(arm,size=0x24) addr:0x020b6d28
-_ZN18PoppingLavaBubblesD0Ev kind:function(arm,size=0x38) addr:0x020b6d4c
-func_ov002_020b6d84 kind:function(arm,size=0x4c) addr:0x020b6d84
-_ZN18PoppingLavaBubbles13InitResourcesEv kind:function(arm,size=0x8) addr:0x020b6dd0
+_ZN11daObjLava_cD1Ev kind:function(arm,size=0x24) addr:0x020b6d28
+_ZN11daObjLava_cD0Ev kind:function(arm,size=0x38) addr:0x020b6d4c
+_ZN11daObjLava_c8BehaviorEv kind:function(arm,size=0x4c) addr:0x020b6d84
+_ZN11daObjLava_c13InitResourcesEv kind:function(arm,size=0x8) addr:0x020b6dd0
 daObjLava_c_classInit kind:function(arm,size=0x30) addr:0x020b6dd8
 _ZN16daObjWaterfall_cD1Ev kind:function(arm,size=0x24) addr:0x020b6e08
 _ZN16daObjWaterfall_cD0Ev kind:function(arm,size=0x38) addr:0x020b6e2c
@@ -1776,7 +1776,6 @@ _ZTI11daObjFire_c kind:data(any) addr:0x02108edc
 _ZTS11daObjFire_c kind:data(any) addr:0x02108ee8
 g_profile_OBJ_RED_FIRE kind:data(any) addr:0x02108ef8
 g_profile_OBJ_BLUE_FIRE kind:data(any) addr:0x02108f14
-_ZTV9BlueFlame kind:data(any) addr:0x02108f38
 _ZTV11daObjFire_c kind:data(any) addr:0x02108f38
 _ZTI17daObjFloatBoard_c kind:data(any) addr:0x02108fb4
 _ZTS17daObjFloatBoard_c kind:data(any) addr:0x02108fc0
@@ -1799,7 +1798,7 @@ _ZTV17daObjKurumajiku_c kind:data(any) addr:0x02109320
 _ZTI11daObjLava_c kind:data(any) addr:0x021093a0
 _ZTS11daObjLava_c kind:data(any) addr:0x021093ac
 g_profile_LAVA kind:data(any) addr:0x021093bc
-_ZTV18PoppingLavaBubbles kind:data(any) addr:0x021093e0
+_ZTV11daObjLava_c kind:data(any) addr:0x021093e0
 _ZTI16daObjWaterfall_c kind:data(any) addr:0x0210945c
 _ZTS16daObjWaterfall_c kind:data(any) addr:0x02109468
 g_profile_WATERFALL kind:data(any) addr:0x0210947c
