+++ config/converted-baseline.json
@@ -889,9 +889,9 @@
     "src/_ZN18NestedHeapIterator6RemoveEP13HeapAllocator.cpp",
     "src/_ZN18NestedHeapIterator8PreviousEP13HeapAllocator.cpp",
     "src/_ZN18NestedHeapIteratorC1Ej.cpp",
-    "src/_ZN18PoppingLavaBubbles13InitResourcesEv.cpp",
-    "src/_ZN18PoppingLavaBubblesD0Ev.cpp",
-    "src/_ZN18PoppingLavaBubblesD1Ev.cpp",
+    "src/actors/daObjLava_c.cpp#_ZN11daObjLava_c13InitResourcesEv",
+    "src/actors/daObjLava_c.cpp#_ZN11daObjLava_cD0Ev",
+    "src/actors/daObjLava_c.cpp#_ZN11daObjLava_cD1Ev",
     "src/_ZN18SolidHeapAllocator5ResetEj.cpp",
     "src/_ZN18TextureTransformer6UpdateER15ModelComponents.cpp",
     "src/_ZN18TextureTransformer7PrepareER8BMD_FileR8BTA_File.cpp",
@@ -1695,10 +1695,10 @@
     "src/_ZN9AnimationD0Ev.cpp",
     "src/_ZN9AnimationD1Ev.cpp",
     "src/_ZN9AnimationD2Ev.cpp",
-    "src/_ZN9BlueFlame13OnTurnIntoEggER6Player.cpp",
-    "src/_ZN9BlueFlame13OnYoshiTryEatEv.cpp",
-    "src/_ZN9BlueFlameD0Ev.cpp",
-    "src/_ZN9BlueFlameD1Ev.cpp",
+    "src/actors/daObjFire_c.cpp#_ZN11daObjFire_c13OnTurnIntoEggER6Player",
+    "src/actors/daObjFire_c.cpp#_ZN11daObjFire_c13OnYoshiTryEatEv",
+    "src/actors/daObjFire_c.cpp#_ZN11daObjFire_cD0Ev",
+    "src/actors/daObjFire_c.cpp#_ZN11daObjFire_cD1Ev",
     "src/_ZN9BootSceneD0Ev.cpp",
     "src/_ZN9BootSceneD1Ev.cpp",
     "src/_ZN9Butterfly16OnPendingDestroyEv.cpp",
@@ -1836,8 +1836,8 @@
     "src/_ZN9dScDSMT_cD0Ev.cpp",
     "src/_ZN9dScDSMT_cD1Ev.cpp",
     "src/_ZN9dThIcon_c6RenderEv.cpp",
-    "src/_ZN9daSCoin_cD0Ev.cpp",
-    "src/_ZN9daSCoin_cD1Ev.cpp",
+    "src/actors/daSCoin_c.cpp#_ZN9daSCoin_cD0Ev",
+    "src/actors/daSCoin_c.cpp#_ZN9daSCoin_cD1Ev",
     "src/_ZN9fLiNdBa_cD1Ev.cpp",
     "src/_ZNK10dBgCh_Actr10IsOnGroundEv.cpp",
     "src/_ZNK10dBgCh_Actr12TouchesWaterEv.cpp",
