+++ include/daObjMarioCap_c.h
@@ -51,7 +51,7 @@
  * The separate 0x021094a0 table-name defect is now fixed: it is configured as
  * _ZTV16daObjWaterfall_c, matching the adjacent retail RTTI. The preceding
  * 0x021093e0 table still belongs to the distinct daObjLava_c actor represented
- * in this tree by PoppingLavaBubbles.
+ * in this tree by daObjLava_c.
  */
 struct daObjMarioCap_c : dEnemyBase_c {
     dCcAc_c  mdCcAc_c;    /* 0x110 */
