+++ include/decl_common.h
@@ -773,7 +773,6 @@ extern int _ZTV8daKrpa_c[];
 extern int _ZTV8daNknk_c[];
 extern int _ZTV8daStar_c[];
 extern int _ZTV8daYegg_c[];
-extern int _ZTV9BlueFlame[];
 extern int _ZTV9Butterfly[];
 extern int _ZTV10daChRoom_c[];
 extern int _ZTV12daDossyCap_c[];
