+++ notes/dScMgBase_c-slots-18-35.md
@@ -114,7 +114,7 @@ Two independent sources agree, and neither of them is this table:
 
 1. **`include/dActor_c.h:131`** declares `virtual int OnYoshiTryEat(); /* slot 18 */`,
    and a dozen further headers repeat the declaration at the same index — BabyPenguin,
-   BlueFlame, BobOmb, BookShot, Coin, Crate, CrazedCrate, `daDossyCap_c`, `daEyBm_c`,
+   daObjFire_c, BobOmb, BookShot, Coin, Crate, CrazedCrate, `daDossyCap_c`, `daEyBm_c`,
    `daKpFr_c`, `daKrb_c`, `daKrpa_c`, `daObjMarioCap_c`, `daTrs_c`. `dActor_c` sits on
    the *other* branch of `fBase_c` (`fBase_c → dBase_c → dActor_c`, where this family
    is `fBase_c → dBase_c → dScene_c → dScMgBase_c`), so on its own this would only be
