+++ notes/data/c-cpp-classification.tsv
@@ -331,7 +331,7 @@ src/Player_ScaleByCharFactor.c	PURE-C					0		None		not-in-delinks
 src/d_a_sanbo_sanbo_body.c	WEAK-refs-only	ov096				1		None		-
 src/d_a_sanbo_sanbo.c	WEAK-refs-only	ov096				1		None		-
 src/d_a_obj_km2_ami_bou.c	WEAK-refs-only	ov045				1		None		-
-src/d_a_obj_lava.c	S1-stores-rom-vptr	ov002			stores _ZTV18PoppingLavaBubbles	1		None	_ZTV18PoppingLavaBubbles	-
+src/d_a_obj_lava.c	S1-stores-rom-vptr	ov002			stores _ZTV11daObjLava_c	1		None	_ZTV11daObjLava_c	-
 src/d_a_obj_power_up_item.c	WEAK-refs-only	ov002				1		None		-
 src/daSCre_c_Spawn.c	S1-stores-rom-vptr	ov018			stores _ZTV8daSCre_c	1		None	_ZTV8daSCre_c	-
 src/d_a_star_star.c	WEAK-refs-only	ov002				1		None		-
@@ -796,7 +796,7 @@ src/_ZN18NestedHeapIterator10FindNestedEPv.c	P3-mangled-method-only	arm9		Nested
 src/_ZN18NestedHeapIterator19RecursiveFindNestedEPv.c	P3-mangled-method-only	arm9		NestedHeapIterator	_ZN18NestedHeapIterator19RecursiveFindNestedEPv	1		True		-
 src/_ZN18NestedHeapIterator5AddAtEP13HeapAllocatorS1_.c	P3-mangled-method-only	arm9		NestedHeapIterator	_ZN18NestedHeapIterator5AddAtEP13HeapAllocatorS1_	1		True		-
 src/_ZN18NestedHeapIteratorC1Ej.cpp	P3-mangled-method-only	arm9		NestedHeapIterator	_ZN18NestedHeapIteratorC1Ej	1		True		-
-src/_ZN18PoppingLavaBubblesD0Ev.c	DTOR-ONLY	ov002	daObjLava_c	PoppingLavaBubbles/daObjLava_c	_ZN18PoppingLavaBubblesD0Ev	1		True	_ZTV18PoppingLavaBubbles	-
+src/_ZN11daObjLava_cD0Ev.c	DTOR-ONLY	ov002	daObjLava_c	daObjLava_c/daObjLava_c	_ZN11daObjLava_cD0Ev	1		True	_ZTV11daObjLava_c	-
 src/_ZN17daObjKm3_Dorifu_cD0Ev.c	DTOR-ONLY	ov047	daObjKm3_Dorifu_c	daObjKm3_Dorifu_c/daObjKm3_Dorifu_c	_ZN17daObjKm3_Dorifu_cD0Ev	1		True	_ZTV10dBgActor_c/_ZTV13daObjDorifu_c/_ZTV17daObjKm3_Dorifu_c	-
 src/_ZN18RotatingPlatformRrD0Ev.c	DTOR-ONLY	ov036	daObjRc_Kaitendai_c	RotatingPlatformRr/daObjRc_Kaitendai_c	_ZN18RotatingPlatformRrD0Ev	1		True	_ZTV10dBgActor_c/_ZTV16daObjKaitendai_c/_ZTV18RotatingPlatformRr	-
 src/_ZN17daObjBk_Ukisima_cD0Ev.c	DTOR-ONLY	ov015	daObjBk_Ukisima_c	daObjBk_Ukisima_c/daObjBk_Ukisima_c	_ZN17daObjBk_Ukisima_cD0Ev	1		True	_ZTV10dBgActor_c/_ZTV16daObjKaitendai_c/_ZTV17daObjBk_Ukisima_c	-
@@ -1028,7 +1028,7 @@ src/_ZN9AnimationD1Ev.cpp	DTOR-ONLY	arm9	dExtFrameCtrl_c	Animation/dExtFrameCtrl
 src/_ZN9AnimationD2Ev.cpp	DTOR-ONLY	arm9		Animation	_ZN9AnimationD2Ev	1		True	_ZTV9Animation	-
 src/_ZN15daObjWc_Obj02_cD0Ev.c	DTOR-ONLY	ov029	daObjWc_Obj02_c	daObjWc_Obj02_c/daObjWc_Obj02_c	_ZN15daObjWc_Obj02_cD0Ev	1		True	_ZTV10dBgActor_c/_ZTV15daObjWc_Obj02_c	-
 src/_ZN15daObjWc_Obj02_cD1Ev.c	DTOR-ONLY	ov029	daObjWc_Obj02_c	daObjWc_Obj02_c/daObjWc_Obj02_c	_ZN15daObjWc_Obj02_cD1Ev	1		True	_ZTV10dBgActor_c/_ZTV15daObjWc_Obj02_c	-
-src/_ZN9BlueFlameD0Ev.c	DTOR-ONLY	ov002	daObjFire_c	BlueFlame/daObjFire_c	_ZN9BlueFlameD0Ev	1		True		-
+src/_ZN11daObjFire_cD0Ev.c	DTOR-ONLY	ov002	daObjFire_c	daObjFire_c/daObjFire_c	_ZN11daObjFire_cD0Ev	1		True		-
 src/_ZN9BootScene13InitResourcesEv.c	P1-vtable+mangled	arm9	dScBoot_c	BootScene/dScBoot_c	vtable dScBoot_c slot 0 @0x2005a58 | _ZN9BootScene13InitResourcesEv	1		True		-
 src/_ZN9BootScene8BehaviorEv.c	P1-vtable+mangled	arm9	dScBoot_c	BootScene/dScBoot_c	vtable dScBoot_c slot 6 @0x2005418 | _ZN9BootScene8BehaviorEv	1		True		-
 src/_ZN9Butterfly16CleanupResourcesEv.c	P1-vtable+mangled	ov100	daBtfly_c	Butterfly/daBtfly_c	vtable daBtfly_c slot 3 @0x2141988 | _ZN9Butterfly16CleanupResourcesEv	1		True		-
