+++ notes/data/class-build-worklist.tsv
@@ -144,7 +144,7 @@ daObjClockHuriko_c	DONE	include\daObjClockHuriko_c.h	dActor_c	0	ov013	6
 daObjClock_c	DONE	include\ClockPaintingHandShort.h  [alias: header uses pre-rename name]	dActor_c	0	ov013	6
 daObjCloset_c	DONE	include\RecRoomCupboard.h  [alias: header uses pre-rename name]	dActor_c	0	ov058	5
 daObjDlPyramidDummy_c	DONE	include\PyramidTag.h  [alias: header uses pre-rename name]	dActor_c	0	ov024	4
-daObjFire_c	DONE	include\BlueFlame.h  [alias: header uses pre-rename name]	dActor_c	0	ov002	6
+daObjFire_c	DONE	include\daObjFire_c.h  [alias: header uses pre-rename name]	dActor_c	0	ov002	6
 daObjFl_Coin_c	PARTIAL	include\daObjFl_Coin_c.h	dActor_c	0	ov064	4
 daObjFlamethrower_c	DONE	include\Flamethrower.h  [alias: header uses pre-rename name]	dActor_c	0	ov095	4
 daObjHeart_c	DONE	include\HealingHeart.h  [alias: header uses pre-rename name]	dActor_c	0	ov002	6
@@ -153,7 +153,7 @@ daObjKb1Billboard_c	DONE	include\daObjKb1Billboard_c.h	dActor_c	0	ov044	6
 daObjKinokoTag_c	DONE	include\daObjKinokoTag_c.h	dActor_c	0	ov002	7
 daObjKm2_Nobiru_c	DONE	include\ExtendingPlatform.h  [alias: header uses pre-rename name]	dActor_c	0	ov045	9
 daObjKumo_c	DONE	include\daObjKumo_c.h	dActor_c	0	ov039	6
-daObjLava_c	DONE	include\PoppingLavaBubbles.h  [alias: header uses pre-rename name]	dActor_c	0	ov002	3
+daObjLava_c	DONE	include\daObjLava_c.h  [alias: header uses pre-rename name]	dActor_c	0	ov002	3
 daObjNumber_c	DONE	include\Number.h  [alias: header uses pre-rename name]	dActor_c	0	ov002	6
 daObjPowerUpItem_c	DONE	include\PowerFlower.h  [alias: header uses pre-rename name]	dActor_c	0	ov002	7
 daObjRc_Hane_c	DONE	include\daObjRc_Hane_c.h	dActor_c	0	ov036	6
