+++ notes/data/tu-merge-candidates.json
@@ -7285,17 +7285,17 @@
    "start": "0x20b5734",
    "end": "0x20b59e0",
    "classes": [
-    "BlueFlame"
+    "daObjFire_c"
    ],
    "n_functions_map": 7,
    "n_files": 7,
    "files": [
-    "src/_ZN9BlueFlameD1Ev.cpp",
-    "src/_ZN9BlueFlameD0Ev.c",
-    "src/_ZN9BlueFlame8BehaviorEv.cpp",
-    "src/_ZN9BlueFlame13InitResourcesEv.cpp",
-    "src/_ZN9BlueFlame13OnTurnIntoEggER6Player.cpp",
-    "src/_ZN9BlueFlame13OnYoshiTryEatEv.cpp",
+    "src/_ZN11daObjFire_cD1Ev.cpp",
+    "src/_ZN11daObjFire_cD0Ev.c",
+    "src/_ZN11daObjFire_c8BehaviorEv.cpp",
+    "src/_ZN11daObjFire_c13InitResourcesEv.cpp",
+    "src/_ZN11daObjFire_c13OnTurnIntoEggER6Player.cpp",
+    "src/_ZN11daObjFire_c13OnYoshiTryEatEv.cpp",
     "src/d_a_obj_fire_obj_blue_fire.c"
    ],
    "ext_mix": {
@@ -15534,16 +15534,16 @@
    "start": "0x20b6d28",
    "end": "0x20b6e08",
    "classes": [
-    "PoppingLavaBubbles",
+    "daObjLava_c",
     "daObjLava_c"
    ],
    "n_functions_map": 5,
    "n_files": 5,
    "files": [
-    "src/_ZN18PoppingLavaBubblesD1Ev.cpp",
-    "src/_ZN18PoppingLavaBubblesD0Ev.c",
+    "src/_ZN11daObjLava_cD1Ev.cpp",
+    "src/_ZN11daObjLava_cD0Ev.c",
     "src/func_ov002_020b6d84.c",
-    "src/_ZN18PoppingLavaBubbles13InitResourcesEv.cpp",
+    "src/_ZN11daObjLava_c13InitResourcesEv.cpp",
     "src/d_a_obj_lava.c"
    ],
    "ext_mix": {
