+++ notes/plan-tu-merge-queue.md
@@ -131,7 +131,7 @@ spent. **Run first; 10 minutes, and it de-risks every later "is it me or is it t
 |[ov002](../config/arm9/overlays/ov002/symbols.txt)| 0x20b05d0|  n=8|  L=126|  `InvisiblePole`|            uncorrob|
 |[ov002](../config/arm9/overlays/ov002/symbols.txt)| 0x20b0748|  n=8|  L=111|  `daCamTag_c`|                uncorrob|
 |[ov002](../config/arm9/overlays/ov002/symbols.txt)| 0x20b07f8|  n=8|  L=121|  `daChRoom_c`|              uncorrob|
-|[ov002](../config/arm9/overlays/ov002/symbols.txt)| 0x20b5734|  n=7|  L=164|  `BlueFlame`|                uncorrob|
+|[ov002](../config/arm9/overlays/ov002/symbols.txt)| 0x20b5734|  n=7|  L=164|  `daObjFire_c`|                uncorrob|
 |[ov002](../config/arm9/overlays/ov002/symbols.txt)| 0x20b6e08|  n=5|  L=67|   `WaterfallMist`|            uncorrob|
 |[ov002](../config/arm9/overlays/ov002/symbols.txt)| 0x20ec388|  n=8|  L=148|  `daWarpkun_c`|             uncorrob|
 |[ov002](../config/arm9/overlays/ov002/symbols.txt)| 0x20f0dd0|  n=8|  L=176|  `OneUpLogo`|                uncorrob|
