--- src/_ZN18PoppingLavaBubbles13InitResourcesEv.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN18PoppingLavaBubbles13InitResourcesEv
-#include "PoppingLavaBubbles.h"
-// recovered name: daObjLava_c_InitResources
-/* recovered: renamed to Class_Method */
-/* daObjLava_c::InitResources - recovered from vtable slot identity */
-s32 PoppingLavaBubbles::InitResources() {
-    return 1;
-}
