--- src/_ZN18PoppingLavaBubblesD1Ev.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN18PoppingLavaBubblesD1Ev
-/* recovered: real C++ destructor -- the compiler emits the whole body.
- * Vtable slot 16: one vtable store, the members in reverse, then ~dActor_c. */
-#include "PoppingLavaBubbles.h"
-
-PoppingLavaBubbles::~PoppingLavaBubbles()
-{
-}
