--- src/_ZN9BlueFlame13OnTurnIntoEggER6Player.cpp
@@ -1,15 +1,0 @@
-//cpp
-// @symbol _ZN9BlueFlame13OnTurnIntoEggER6Player
-// recovered name: BlueFlame_OnTurnIntoEgg
-/* daObjFire_c::OnTurnIntoEgg -- vtable slot 19, verified against ov002
- * relocs.txt: _ZTV9BlueFlame (0x02108f38) + 0x4c -> 0x020b599c, exactly this
- * placeholder's former address (former name func_ov002_020b599c). The ROM
- * body is a single `bx lr` -- an intentional no-op override, not a stub.
- * Matched byte-for-byte with mwccarm 2004/b56 (ov002).
- */
-#include "BlueFlame.h"
-#include "Player.h"
-
-int BlueFlame::OnTurnIntoEgg(Player &player)
-{
-}
