--- src/_ZN9BlueFlame13OnYoshiTryEatEv.cpp
@@ -1,7 +1,0 @@
-//cpp
-// @symbol _ZN9BlueFlame13OnYoshiTryEatEv
-#include "BlueFlame.h"
-/* recovered: renamed to Class_Method */
-s32 BlueFlame::OnYoshiTryEat() {
-    return 5;
-}
