--- src/_ZN9BlueFlameD1Ev.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN9BlueFlameD1Ev
-/* recovered: real C++ destructor -- the compiler emits the whole body.
- * Vtable slot 16: one vtable store, the members in reverse, then ~dActor_c. */
-#include "BlueFlame.h"
-
-BlueFlame::~BlueFlame()
-{
-}
