--- src/d_a_obj_fire_obj_red_fire.c
@@ -1,22 +1,0 @@
-// @symbol daObjFire_c_classInit_OBJ_RED_FIRE
-/* recovered: vtable identified, globals resolved, declarations from a shared header */
-#include "decl_Actor.h"
-#include "decl_ActorBase.h"
-#include "decl_dCcAc_c.h"
-#include "decl_common.h"
-/* recovered: vtable identified, globals resolved */
-/* resolved: VT0 = _ZTV9BlueFlame */
-/* Reconstructed source-style name: SM64DS proves daObjFire_c through RTTI,
- * allocation size, vtable identity, and the OBJ_RED_FIRE registry profile;
- * later EAD lineage supplies classInit. Exact original spelling is not
- * preserved. Historical alias: RedFlame_Spawn. */
-int *daObjFire_c_classInit_OBJ_RED_FIRE(void)
-{
-    int *p = (int *)_ZN7fBase_cnwEj(280);
-    if (p) {
-        _ZN8dActor_cC2Ev(p);
-        p[0] = (int)_ZTV9BlueFlame;
-        _ZN7dCcAc_cC1Ev((char *)p + 0xe4);
-    }
-    return p;
-}
