+++ symbols/actor_renames.tsv
@@ -3454,3 +3454,15 @@ ov006	0x020f300c	func_ov006_020f300c	_ZN12dScMgLuigi_c10StateCatchEv	coined memb
 ov006	0x020f319c	func_ov006_020f319c	_ZN12dScMgLuigi_c9StatePlayEv	coined member name; member-ness is ROM-proven: the ov006 .data word at 0x0213ce44 is 0x020f319c and the word after it is 0, which is how mwccarm spells a pointer-to-member-function constant with a zero this-adjustment, and the dispatch sites in this TU that reach it through a pointer-to-member on this object byte-match the cartridge (tools/tubuild.py verify ov006/dScMgLuigi_c, 58/58). The NAME is coined -- the cartridge carries this address and no identifier -- and describes the matched body only: the arm of the Behavior state machine that runs while +0x4f78 is 2 (see StateCatch's row for how that mapping is derived) -- it ticks the gauge and, unless the reveal has already started, moves the slots, tests the touch and runs the penalty popups
 ov006	0x020f31dc	func_ov006_020f31dc	_ZN12dScMgLuigi_c10StatePlaceEv	coined member name; member-ness is ROM-proven: the ov006 .data word at 0x0213ce24 is 0x020f31dc and the word after it is 0, which is how mwccarm spells a pointer-to-member-function constant with a zero this-adjustment, and the dispatch sites in this TU that reach it through a pointer-to-member on this object byte-match the cartridge (tools/tubuild.py verify ov006/dScMgLuigi_c, 58/58). The NAME is coined -- the cartridge carries this address and no identifier -- and describes the matched body only: the arm of the Behavior state machine that runs while +0x4f78 is 1 (see StateCatch's row) -- it places eight more slots a frame and moves to state 2 once the board is full and both curtain entries are idle
 ov006	0x020f3260	func_ov006_020f3260	_ZN12dScMgLuigi_c10StateSetupEv	coined member name; member-ness is ROM-proven: the ov006 .data word at 0x0213cdd4 is 0x020f3260 and the word after it is 0, which is how mwccarm spells a pointer-to-member-function constant with a zero this-adjustment, and the dispatch sites in this TU that reach it through a pointer-to-member on this object byte-match the cartridge (tools/tubuild.py verify ov006/dScMgLuigi_c, 58/58). The NAME is coined -- the cartridge carries this address and no identifier -- and describes the matched body only: the arm of the Behavior state machine that runs while +0x4f78 is 0 (see StateCatch's row) -- it hides the reveal icon, resets the curtain entries, moves to state 1 and picks the target
+ov002	0x020b6d28	_ZN18PoppingLavaBubblesD1Ev	_ZN11daObjLava_cD1Ev	vtable slot 16; class renamed to its ROM RTTI name 11daObjLava_c (_ZTS at 0x021093ac); coined PoppingLavaBubbles retired
+ov002	0x020b6d4c	_ZN18PoppingLavaBubblesD0Ev	_ZN11daObjLava_cD0Ev	vtable slot 17; class renamed to its ROM RTTI name 11daObjLava_c; coined PoppingLavaBubbles retired
+ov002	0x020b6d84	func_ov002_020b6d84	_ZN11daObjLava_c8BehaviorEv	vtable slot 6 of _ZTV11daObjLava_c holds 0x020b6d84 (own override, not the dActor_c word)
+ov002	0x020b6dd0	_ZN18PoppingLavaBubbles13InitResourcesEv	_ZN11daObjLava_c13InitResourcesEv	vtable slot 0; class renamed to its ROM RTTI name 11daObjLava_c; coined PoppingLavaBubbles retired
+ov002	0x021093e0	_ZTV18PoppingLavaBubbles	_ZTV11daObjLava_c	vtable address point; class renamed to its ROM RTTI name 11daObjLava_c; coined PoppingLavaBubbles retired
+ov002	0x020b5734	_ZN9BlueFlameD1Ev	_ZN11daObjFire_cD1Ev	vtable slot 16; class renamed to its ROM RTTI name 11daObjFire_c (_ZTS at 0x02108ee8); coined BlueFlame retired
+ov002	0x020b5764	_ZN9BlueFlameD0Ev	_ZN11daObjFire_cD0Ev	vtable slot 17; class renamed to its ROM RTTI name 11daObjFire_c; coined BlueFlame retired
+ov002	0x020b57a8	_ZN9BlueFlame8BehaviorEv	_ZN11daObjFire_c8BehaviorEv	vtable slot 6; class renamed to its ROM RTTI name 11daObjFire_c; coined BlueFlame retired
+ov002	0x020b5940	_ZN9BlueFlame13InitResourcesEv	_ZN11daObjFire_c13InitResourcesEv	vtable slot 0; class renamed to its ROM RTTI name 11daObjFire_c; coined BlueFlame retired
+ov002	0x020b599c	_ZN9BlueFlame13OnTurnIntoEggER6Player	_ZN11daObjFire_c13OnTurnIntoEggER6Player	vtable slot 19; class renamed to its ROM RTTI name 11daObjFire_c; the R6Player reference spelling is a coined guess the bytes cannot distinguish from a pointer
+ov002	0x020b59a0	_ZN9BlueFlame13OnYoshiTryEatEv	_ZN11daObjFire_c13OnYoshiTryEatEv	vtable slot 18; class renamed to its ROM RTTI name 11daObjFire_c; coined BlueFlame retired
+ov002	0x02108f38	_ZTV9BlueFlame	_ZTV11daObjFire_c	coined alias row deleted; the address already carried the ROM RTTI vtable name
